copyedits
L 46: I suggest to add an example or reference on parentheses sometimes not being a reliable indicator of original combination.

L70: Looking forward to the details.

L 78: Is there a Bufonidae equivalent to the Li et al. 2009 paper on phylogenetic revision of Rhacophoridae? If so, I would include it.

L 100: Perhaps round to 2 significant digits?